250624 controller 매핑 수정, import/param 수정, jpa 자료형, 오타 수정, build.gradle에서 spring security 주석 제거, postgreSQL 추가

+++ src/main/java/com/team/arium/domain/Cnsl_Aply.java
@@ -20,7 +20,7 @@ public class Cnsl_Aply {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "cnsl_aply_id")
-    private Long cnslAplyId;
+    private Integer cnslAplyId;
     
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "std_id", nullable = false)
